revert navigator server to bt_navigator and remame bt_action_navigator
