+++ nav2_bt_navigator/include/nav2_bt_navigator/navigators/navigate_through_poses.hpp
@@ -21,7 +21,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_action/rclcpp_action.hpp"
 #include "geometry_msgs/msg/pose_stamped.hpp"
-#include "nav2_bt_navigator/bt_navigator.hpp"
+#include "nav2_bt_navigator/bt_action_navigator.hpp"
 #include "nav2_msgs/action/navigate_through_poses.hpp"
 #include "nav_msgs/msg/path.hpp"
 #include "nav2_util/robot_utils.hpp"
@@ -36,7 +36,7 @@ namespace nav2_bt_navigator
  * @brief A navigator for navigating to a bunch of intermediary poses
  */
 class NavigateThroughPosesNavigator
-  : public nav2_bt_navigator::BTNavigator<nav2_msgs::action::NavigateThroughPoses>
+  : public nav2_bt_navigator::BtActionNavigator<nav2_msgs::action::NavigateThroughPoses>
 {
 public:
   using ActionT = nav2_msgs::action::NavigateThroughPoses;
@@ -46,7 +46,7 @@ public:
    * @brief A constructor for NavigateThroughPosesNavigator
    */
   NavigateThroughPosesNavigator()
-  : BTNavigator() {}
+  : BtActionNavigator() {}
 
   /**
    * @brief Get action name for this navigator
